+++ src/steps/index.ts
@@ -1,6 +1,11 @@
 import { accessSteps } from './access';
 import { organizationSteps } from './account';
+import { findingSteps } from './finding';
 
-const integrationSteps = [...organizationSteps, ...accessSteps];
+const integrationSteps = [
+  ...organizationSteps,
+  ...accessSteps,
+  ...findingSteps,
+];
 
 export { integrationSteps };